refactor: extract spectrogram and FFT from plots.py (#712)
Move PeakSampleScan scan layer, FFT spectrum aggregation, and 2-D
spectrogram builders into diagnostics/spectrogram.py (313 lines).

plots.py shrinks from 752 to 474 lines, keeping only series
extraction, peak-table ranking, and the _plot_data orchestrator.

Test imports updated to point at the new module.
